+++ CHANGELOG.md
@@ -6,8 +6,15 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Semantic Versioning].
 
 <!-- references -->
-[Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
-[Semantic Versioning]: https://semver.org/spec/v2.0.0.html
+
+[keep a changelog]: https://keepachangelog.com/en/1.0.0/
+[semantic versioning]: https://semver.org/spec/v2.0.0.html
+
+## [0.5.1] - 2022-08-20
+
+### Added
+
+- Add `envelopespec.Envelope.SourceSite`
 
 ## [0.5.0] - 2021-01-20
 
@@ -57,7 +64,8 @@ The format is based on [Keep a Changelog], and this project adheres to
 - Initial release
 
 <!-- references -->
-[Unreleased]: https://github.com/dogmatiq/interopspec
+
+[unreleased]: https://github.com/dogmatiq/interopspec
 [0.1.0]: https://github.com/dogmatiq/interopspec/releases/tag/v0.1.0
 [0.2.0]: https://github.com/dogmatiq/interopspec/releases/tag/v0.2.0
 [0.2.1]: https://github.com/dogmatiq/interopspec/releases/tag/v0.2.1